==> notebooks/1_alizadeh_preprocessing.ipynb <==
import pandas as pd

df = pd.read_excel('../data/classification/alizadeh.xlsx')

df.head()

df.isnull().sum() 

# Заполнение не требуется

mapping_sex = {'Male':1,'Fmale':2}
df['Sex'] = df['Sex'].map(mapping_sex)


mapping_yes_no = {'N':0,'Y':1}
df['Obesity'] = df['Obesity'].map(mapping_yes_no)
df['CRF'] = df['CRF'].map(mapping_yes_no)
df['CVA'] = df['CVA'].map(mapping_yes_no)
df['Airway disease'] = df['Airway disease'].map(mapping_yes_no)
df['Thyroid Disease'] = df['Thyroid Disease'].map(mapping_yes_no)
df['CHF'] = df['CHF'].map(mapping_yes_no)
df['DLP'] = df['DLP'].map(mapping_yes_no)
df['Weak Peripheral Pulse'] = df['Weak Peripheral Pulse'].map(mapping_yes_no)
df['Lung rales'] = df['Lung rales'].map(mapping_yes_no)
df['Systolic Murmur'] = df['Systolic Murmur'].map(mapping_yes_no)
df['Diastolic Murmur'] = df['Diastolic Murmur'].map(mapping_yes_no)
df['Dyspnea'] = df['Dyspnea'].map(mapping_yes_no)
df['Atypical'] = df['Atypical'].map(mapping_yes_no)
df['Nonanginal'] = df['Nonanginal'].map(mapping_yes_no)
df['Exertional CP'] = df['Exertional CP'].map(mapping_yes_no)
df['LowTH Ang'] = df['LowTH Ang'].map(mapping_yes_no)
df['LVH'] = df['LVH'].map(mapping_yes_no)
df['Poor R Progression'] = df['Poor R Progression'].map(mapping_yes_no)

mapping_BBB = {'N':0,'LBBB':1,'RBBB':2}
df['BBB'] = df['BBB'].map(mapping_BBB)

mapping_VHD = {'mild':0,'Moderate':1,'N':2,'Severe':3}
df['VHD'] = df['VHD'].map(mapping_VHD)

mapping_normal_stenotic = {'Normal':0,'Stenotic':1}
df['LAD'] = df['LAD'].map(mapping_normal_stenotic)
df['LCX'] = df['LCX'].map(mapping_normal_stenotic)
df['RCA'] = df['RCA'].map(mapping_normal_stenotic)

mapping_cath = {'CAD':0,'Normal':1}
df['Cath'] = df['Cath'].map(mapping_cath)
df

pd.get_dummies(df).to_csv("../data/classification/alizadeh_preprocessed.csv", index_label = False) 

df1 = pd.read_csv("../data/classification/alizadeh_preprocessed.csv", sep=',')
df1.head()
