==> notebooks/1_facebook_preprocessing.ipynb <==
import pandas as pd

df = pd.read_csv("../data/regression/facebook.csv", sep=';')

df.head()

df.info()

df.isnull().sum()

df['Paid'].fillna(0, inplace=True)
df['like'].fillna(df['like'].median(), inplace=True)
df['share'].fillna(df['share'].mean(), inplace=True)

df['Paid'] = df.Paid.astype(int)
df['like'] = df.like.astype(int)
df['share'] = df.share.astype(int)

df.isnull().sum()

df.mean()

df.dropna()

pd.get_dummies(df, prefix='Type', columns=['Type'])

pd.get_dummies(df)

pd.get_dummies(df).to_csv("../data/regression/facebook_preprocessed.csv", index_label = False)

df1 = pd.read_csv("../data/regression/facebook_preprocessed.csv", sep=',')

df1.head()
