==> notebooks/3_regression.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LinearRegression
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
from sklearn.tree import DecisionTreeRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import Lasso
from sklearn.linear_model import Ridge
from sklearn.linear_model import ElasticNet

df = pd.read_csv("../data/regression/facebook_preprocessed.csv") 

df.head()

df.info()

y = df["Page total likes"]
y

X = df[df.columns[2:]]
X

# Разбиение выборки

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)

scaler = StandardScaler()
scaler.fit(X_train)

X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)

# Линейная регрессия

from sklearn.linear_model import LinearRegression
from sklearn.metrics import r2_score, mean_absolute_error, mean_squared_error
from math import sqrt

lr = LinearRegression()
lr.fit(X_train, y_train)
y_pred_lr = lr.predict(X_test) 

print(r2_score(y_test, y_pred_lr))
print(mean_squared_error(y_test, y_pred_lr))
print(mean_squared_error(y_test, y_pred_lr, squared=False))
print(mean_absolute_error(y_test, y_pred_lr))


# Регрессия дерева решений

dtr = DecisionTreeRegressor()
dtr.fit(X_train, y_train)
y_pred_dtr = dtr.predict(X_test)

print(r2_score(y_test, y_pred_dtr))
print(mean_squared_error(y_test, y_pred_dtr))
print(mean_squared_error(y_test, y_pred_dtr, squared=False))
print(mean_absolute_error(y_test, y_pred_dtr))

  dtr.get_params().keys()

dtr_parameters_grid = {'criterion' : ['mse', 'friedman_mse', 'mae', 'poisson'],
                       'splitter' : ['best', 'random'],
                       'max_depth' : range(1, 11),                  
}

dtr_grid = GridSearchCV(dtr, dtr_parameters_grid, cv=5, n_jobs=-1, verbose=True)
dtr_grid.fit(X_train, y_train)
dtr_grid.best_params_

dtr_g = DecisionTreeRegressor(criterion='mse', max_depth=4,splitter='best')
dtr_g.fit(X_train, y_train)
y_pred_dtr_g = dtr_g.predict(X_test)

print(r2_score(y_test, y_pred_dtr_g))
print(mean_squared_error(y_test, y_pred_dtr_g))
print(mean_squared_error(y_test, y_pred_dtr_g, squared=False))
print(mean_absolute_error(y_test, y_pred_dtr_g))

# LASSO

lasso = Lasso(alpha=1.0, max_iter=10000)
lasso.fit(X_train, y_train)
y_pred_lasso = lasso.predict(X_test)

print(r2_score(y_test, y_pred_lasso))
print(mean_squared_error(y_test, y_pred_lasso))
print(mean_squared_error(y_test, y_pred_lasso, squared=False))
print(mean_absolute_error(y_test, y_pred_lasso))

 lasso.get_params().keys()

lasso_parameters_grid = {'alpha' : np.arange(0.01, 1.0, 0.05),
                         'normalize' : [True, False],
                         'max_iter' : [10000],
                         'selection' : ['cyclic', 'random'],
                         'warm_start':[True, False]
                         }

lasso_grid = GridSearchCV(lasso, lasso_parameters_grid, cv = 5, n_jobs=-1, verbose=True)
lasso_grid.fit(X_train, y_train)
lasso_grid.best_params_

 lasso_g = Lasso(alpha= 0.96, max_iter=10000, normalize=True, selection='random', warm_start=True)
lasso_g.fit(X_train,y_train)
y_pred_lasso_g = lasso_g.predict(X_test)

print(r2_score(y_test, y_pred_lasso_g))
print(mean_squared_error(y_test, y_pred_lasso_g))
print(mean_squared_error(y_test, y_pred_lasso_g, squared=False))
print(mean_absolute_error(y_test, y_pred_lasso_g))

# Гребневая регрессия

ridge = Ridge()
ridge.fit(X_train,y_train)
y_pred_ridge = ridge.predict(X_test)

print(r2_score(y_test, y_pred_ridge))
print(mean_squared_error(y_test, y_pred_ridge))
print(mean_squared_error(y_test, y_pred_ridge, squared=False))
print(mean_absolute_error(y_test, y_pred_ridge))

ridge.get_params().keys()

ridge_parameters_grid = {'alpha' : np.arange(0.01, 1.0, 0.05),
                         'max_iter' : [10000],
                         'solver' : ['auto', 'svd', 'cholesky', 'lsqr', 'sparse_cg', 'sag', 'saga'],
                         'fit_intercept' : [True, False]
                         }

ridge_grid = GridSearchCV(ridge, ridge_parameters_grid, cv = 5, n_jobs=-1, verbose=True)
ridge_grid.fit(X_train, y_train)
ridge_grid.best_params_

ridge_g = Ridge(alpha=0.96, fit_intercept=True, max_iter=10000, solver='saga')
ridge_g.fit(X_train,y_train)
y_pred_ridge_g = ridge_g.predict(X_test)

print(r2_score(y_test, y_pred_ridge_g))
print(mean_squared_error(y_test, y_pred_ridge_g))
print(mean_squared_error(y_test, y_pred_ridge_g, squared=False))
print(mean_absolute_error(y_test, y_pred_ridge_g))

# Elastic Net регрессия

en = ElasticNet()
en.fit(X_train, y_train)
y_pred_en = en.predict(X_test)

print(r2_score(y_test, y_pred_en))
print(mean_squared_error(y_test, y_pred_en))
print(mean_squared_error(y_test, y_pred_en, squared=False))
print(mean_absolute_error(y_test, y_pred_en))

en.get_params().keys()

en_parameters_grid = {'alpha' : np.arange(0.01, 1.0, 0.05),
                      'l1_ratio' : np.arange(0.1, 1.0, 0.1),
                      'max_iter' : [10000],
                      'fit_intercept' : [True, False],
                      'normalize' : [True, False],
                      'selection' : ['cyclic', 'random'],
                     }

en_grid = GridSearchCV(en, en_parameters_grid, cv = 5, n_jobs=-1, verbose=True)
en_grid.fit(X_train, y_train)
en_grid.best_params_

en_g = ElasticNet(alpha=0.81, fit_intercept=True, l1_ratio=0.9, max_iter=10000, normalize=False, selection='cyclic')
en_g.fit(X_train, y_train)
y_pred_en_g = en_g.predict(X_test)

print(r2_score(y_test, y_pred_en_g))
print(mean_squared_error(y_test, y_pred_en_g))
print(mean_squared_error(y_test, y_pred_en_g, squared=False))
print(mean_absolute_error(y_test, y_pred_en_g))
