==> notebooks/7_keras.ipynb <==
import pandas as pd
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from keras.models import Sequential
from keras.layers import Dense
from keras.utils.vis_utils import plot_model

# Решение задачи классификации с помощью Keras

##  Подготовка данных

df = pd.read_table('../data/classification/alizadeh_preprocessed.csv', sep=',')

df.head()

X = df.iloc[:,:26].values
y = df.iloc[:,27].values

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)

Объявляем StandardScaler для нормализации данных.

scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

##  Модель

model = Sequential()
model.add(Dense(26, input_dim=26, activation='relu'))
model.add(Dense(1, activation='sigmoid'))
model.summary()

model.compile(loss='mean_squared_error', optimizer='adam', metrics=['accuracy'])

model.fit(X_train, y_train, epochs=250)

 _, accuracy = model.evaluate(X, y)
print('Accuracy: %.2f' % (accuracy*100))

 model.save('model')
